Remove duplicated test
